neigbor list 3/3 fix - added long long to prevent overflof

+++ number_of_complements_edges.h
@@ -2,7 +2,7 @@
 
 #include "graph.h"
 
-unsigned int get_edge_count(graph_t * graph);
+long long int get_edge_count(graph_t * graph);
 
 void number_of_complements_edges(graph_t * graph);
